Generate individual dependency files
This removes Makefile.depends. Instead, '.d' files are included from the 'dep'
directory. This speeds up building because dependencies for every file don't
need to be regenerated if only one changes.

As a bonus, dependencies also won't be generated if only clean type targets
are going to be run. Also added a 'distclean' target, which cleans both
objects and dependency files.