Ensure the servername fits into MAXSERVERNAME before setting it
This commit adds a verification before setting the servername console
variable. We now check that it fits within the MAXSERVERNAME length and
we cancel setting it if it doesn't.
Letting the servername be more than MAXSERVERNAME could lead to crashes
when trying to edit the server name from the menu so, we now avoid
those.